requirements.txt
beautifulsoup4==4.6.0
gensim==4.1.0
ipywidgets
joblib==1.0.1
JPype1==1.2.1
konlpy==0.5.2
lxml==4.6.3
matplotlib==3.4.1
nbresuse==0.4.0
nltk==3.6.2
numpy==1.19.5
pandas==1.2.3
pandocfilters==1.4.3
parmap==1.5.2
psutil==5.8.0
psycopg2-binary==2.8.6
pytrends==4.7.3
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
sklearn==0.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow-addons==0.12.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
tqdm==4.59.0
tweepy==3.10.0
urllib3==1.26.4
